refactor: removed vlite database class
